Update UI for BrowserSearch, NetworkSetting screen, update FlatListScreen
